% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a2d_circ.r
\name{a2d_circ}
\alias{a2d_circ}
\title{Compute diameter of a circle from its area}
\usage{
a2d_circ(A)
}
\arguments{
\item{A}{area of a circle}
}
\description{
Compute diameter of a circle from its area
}
\examples{
# Example
x <- rnorm(10)
y <- rnorm(10)
n <- sample(1:100,size=10,replace=TRUE)
cex <- 3
cex.scaled <- a2d_circ(cex*(n/max(n))) # Scale cex (diameter) of points so area is proportional to n
plot(x,y,cex=cex)
plot(x,y,cex=cex.scaled)

}
\author{
Nikolai Klibansky
}
