% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_spawn.r
\name{calc_spawn}
\alias{calc_spawn}
\title{Calculate number of batches produced by each spawning capable female}
\usage{
calc_spawn(
  date,
  y,
  nb_calc = FALSE,
  SC_pd = NA,
  dh_SI = 24,
  fd1 = "1970-01-01",
  time_unit = "day",
  print_result = FALSE
)
}
\arguments{
\item{date}{A vector of dates as a factor with numeric levels (e.g. "001", "365")}

\item{y}{A vector of spawning indicator (SI) presence/absence as a binary variable}

\item{nb_calc}{logical. Indicate if batch number calculations should be made}

\item{SC_pd}{Limits of main SC period. NEEDED ONLY for batch number calculations}

\item{dh_SI}{Estimated duration of spawning indicator in hours (defaults to 24 hours)}

\item{fd1}{First date of the fiscal year (or calendar year is all dates occur within it). Defaults to the first date of 1970 (R standard origin date)}

\item{time_unit}{unit of time to use for grouping observations for plotting proportions defaults to "day", but could also be "week" or "month"}

\item{print_result}{logical. Indicate whether or not print out results. defaults to FALSE}
}
\description{
Calculate number of batches produced by each spawning capable female using standard calculations.
}
\examples{
\dontrun{
# Simulate spawning with defaults
set.seed(23456)
out_spawn <- sim_spawn()


# Simulate sampling with defaults
out_sample <- sim_sample(sim_spawn_out = out_spawn)

# plot results with defaults
plot_spawn(data_pop=out_spawn$data_pop, data_smp=out_spawn$data_smp)

# Apply standard calculations to compute batch number and related quantities.
with(out_sample$data_smp,
     {
       calc_spawn(date=date, y=SC, dh_SI=24,nb_calc=TRUE,print_result = TRUE)
     })
}

}
\author{
Nikolai Klibansky
}
