% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceilingf.r
\name{ceilingf}
\alias{ceilingf}
\title{Compute the greatest multiple of a factor f which is greater than or equal to a value x}
\usage{
ceilingf(x, f)
}
\arguments{
\item{x}{numeric vector}

\item{f}{factor}
}
\description{
Compute the greatest multiple of a factor f which is greater than or equal to a value x
}
\examples{
\dontrun{
# Example
}

}
\author{
Nikolai Klibansky
}
