% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covmat_SI_L}
\alias{covmat_SI_L}
\title{Variance-covariance matrix for parameters in four parameter logistic model fit to red snapper spawning indicator-size data.}
\format{
\subsection{\code{covmat_SI_L}}{

\describe{
\item{m}{Parameter m. point of inflection}
\item{s}{Parameter s. scale parameter}
\item{a}{Parameter a. upper asymptote}
\item{f}{Parameter f. lower asymptote}
}
}
}
\source{
\url{https://academic.oup.com/mcf/article/7/1/233/7827111?guestAccessKey=}
}
\usage{
covmat_SI_L
}
\description{
Covariance matrix for parameters in four parameter logistic model fit \eqn{y=f+(a-f)/(1+exp(-(x-m)/s))} to red snapper spawning indicator-size data stored in data_SI_L. The logistic model was fit with with \code{\link[stats]{optim}}, minimizing the negative log-likelihood. Running \code{\link[base]{solve}} on the hessian matrix returns the variance-covariance matrix among the model parameters.
}
\keyword{datasets}
