% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_fb_L}
\alias{data_fb_L}
\title{Red snapper batch fecundity-length data}
\format{
\subsection{\code{data_fb_L}}{

\describe{
\item{Source}{Agency which collected the data (FWRI = Florida Fish and Wildlife Research Institute, SCDNR = South Carolina Department of Natural Resources)}
\item{Date}{Date (MM/DD/YYYY) of data collection}
\item{Month}{Month of data collection}
\item{TL}{Fish total length (mm)}
\item{TW}{Fish total weight (g)}
\item{SW}{Fish somatic weight (g)}
\item{Age}{Fish age (yr)}
\item{fb}{Fish batch fecundity (eggs)}
}
}
}
\source{
\url{https://sedarweb.org/documents/s41dw49-estimates-of-reproductive-activity-in-red-snapper-by-size-season-and-time-of-day-with-nonlinear-models/}
}
\usage{
data_fb_L
}
\description{
Batch fecundity, length, and related biological data, provided to the 2017 (SEDAR 41) stock assessment of the Southeast US Atlantic stock of red snapper.
}
\keyword{datasets}
