% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_SI_d}
\alias{data_SI_d}
\title{Seasonal patterns in red snapper spawning activity}
\format{
\subsection{\code{data_SI_d}}{

\describe{
\item{Period}{Numeric ID of sampling period, in sequence.}
\item{Start.date}{Start date of time period}
\item{N.dates}{Number of sampling dates in each time period}
\item{N.f}{Number of females sampled}
\item{N.f.SI}{Number of females sampled with spawning indicators}
\item{P.SI}{Proportion of females with spawing indicators}
\item{Age.mn}{Mean age of females}
\item{Age.rng}{Age range of females}
}
}
}
\source{
\url{https://sedarweb.org/documents/sedar31-dw07-expanded-annual-stock-assessment-survey-2011-red-snapper-reproduction/}
}
\usage{
data_SI_d
}
\description{
Data shows the proportions of female red snapper with spawning indicators by time of year, from Table 1 of: Fitzhugh, G.R., E.T. Lang, and H. Lyon. 2012. Expanded Annual Stock Assessment Survey 2011: Red Snapper Reproduction. SEDAR31-DW07. SEDAR, North Charleston, SC. 33 pp.
}
\keyword{datasets}
