% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_SI_d_bin}
\alias{data_SI_d_bin}
\title{Seasonal patterns in red snapper spawning activity (raw version)}
\format{
\subsection{\code{data_SI_d_bin}}{

\describe{
\item{Start.date}{Start date of time period}
\item{SI}{Presence (1) or absence (0) of spawing indicators}
}
}
}
\source{
\url{https://sedarweb.org/documents/sedar31-dw07-expanded-annual-stock-assessment-survey-2011-red-snapper-reproduction/}
}
\usage{
data_SI_d_bin
}
\description{
Data converted from \code{data_SI_d} so that each row represents data for one fish and SI is binary
}
\keyword{datasets}
