% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_SI_h}
\alias{data_SI_h}
\title{Diel patterns in red snapper spawning activity}
\format{
\subsection{\code{data_SI_h}}{

\describe{
\item{CruiseNum}{Numeric ID of sampling cruise, in sequence.}
\item{Cruise}{Alphanumeric cruise ID}
\item{Hour}{Hour of sampling}
\item{Fmat}{Number of mature females sampled}
\item{HO}{Number of females with hydrated oocytes}
\item{POF}{Number of females with post-ovulatory follicles}
}
}
}
\source{
\url{https://onlinelibrary.wiley.com/doi/abs/10.1111/j.0022-1112.2006.00951.x}
}
\usage{
data_SI_h
}
\description{
Data shows the proportions of female red snapper with spawning indicators by time of day, from data captured from Figure 1b of: Jackson, M. W., Nieland, D. L., & Cowan Jr, J. H. (2006). Diel spawning periodicity of red snapper Lutjanus campechanus in the northern Gulf of Mexico. Journal of Fish Biology, 68(3), 695-706.
}
\keyword{datasets}
