% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_SI_h_bin}
\alias{data_SI_h_bin}
\title{Diel patterns in red snapper spawning activity (raw version)}
\format{
\subsection{\code{data_SI_h_bin}}{

\describe{
\item{CruiseNum}{Numeric ID of sampling cruise, in sequence.}
\item{Cruise}{Alphanumeric cruise ID}
\item{Hour}{Hour of sampling}
\item{Fmat}{Number of mature females sampled}
\item{HO}{Presence (1) or absence (0) of hydrated oocytes}
\item{POF}{Presence (1) or absence (0) of post-ovulatory follicles}
\item{Hour.shf}{Shifted hour of sampling, used for plotting distribution of POF. Hours 0-8 are shifted by +24 hours.}
}
}
}
\source{
\url{https://onlinelibrary.wiley.com/doi/abs/10.1111/j.0022-1112.2006.00951.x}
}
\usage{
data_SI_h_bin
}
\description{
Data converted from \code{data_SI_h} so that each row represents data for one fish and SI is binary
}
\keyword{datasets}
