% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_SI_L}
\alias{data_SI_L}
\title{Red snapper spawning indicator-size data}
\format{
\subsection{\code{data_SI_L}}{

\describe{
\item{Age}{Age of females}
\item{N.f}{Number of females sampled}
\item{P.SI}{Proportion of females with spawing indicators}
\item{P.SI.adj}{Proportion of females with spawing indicators, adjusted by multiplying by a correction factor (CF) based on a spawning indicator duration of 34 hours \eqn{CF = (10+24)/24} }
\item{N.f.SI}{Number of females sampled with spawning indicators}
\item{L}{Fish length (mm) estimated from ages. \eqn{L = 848.4797*(1-exp(-0.2188*(Age-(-0.0611))))]}. SEDAR 31 2.9.1 pg 32 Eq 4. converted to mm and g by B0'=B0*(0.0393701^B1)/0.00220462}
}
}
}
\source{
\url{https://academic.oup.com/mcf/article/7/1/233/7827111?guestAccessKey=}
}
\usage{
data_SI_L
}
\description{
Red snapper spawning indicator-size data, from an earlier draft of: Porch, C. E., Fitzhugh, G. R., Lang, E. T., Lyon, H. M., & Linton, B. C. (2015). Estimating the dependence of spawning frequency on size and age in Gulf of Mexico Red Snapper. Marine and Coastal Fisheries, 7(1), 233-245. Similar to Table 1 of the published manuscript.
}
\keyword{datasets}
