% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diel_sample.r
\name{diel_sample}
\alias{diel_sample}
\title{Diel sampling}
\usage{
diel_sample(n, freq = 1, h_sunrise = 6, h_sunset = 17, pattern = "day")
}
\arguments{
\item{n}{number of days to draw hours for}

\item{freq}{frequency of sampling (e.g. every 1 hour, every 2 hours)}

\item{h_sunrise}{hour of sunrise}

\item{h_sunset}{hour of sunset}

\item{pattern}{sample pattern: "rand"= randomly chooses whether to sample during the AM or PM, "alt"= alternate periods (e.g. 1,2,1,2,1,2,..), "all"= sample every hour of the day, "day"=sample every hour during daylight hours (e.g. 0600-1700h), "day_rand"=sample only during daylight hours and randomly choose 12 sampling times (i.e. repeat some hours and skip some hours)}
}
\value{
Character vector of times of day in 24h format: 00:00 to 23:00
}
\description{
A function which randomly chooses hours from the first or second half of the day
}
\examples{
diel_sample(n=1,pattern="day")
diel_sample(n=1,pattern="rand")
diel_sample(n=1,pattern="alt")
diel_sample(n=1,pattern="day_rand")
diel_sample(n=1,freq=1,pattern="all")
diel_sample(n=1,freq=2,pattern="all")
diel_sample(n=1,freq=4,pattern="all")
}
\author{
Nikolai Klibansky
}
