% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtnorm2.r
\name{dtnorm2}
\alias{dtnorm2}
\title{Probability density of a variable following a truncated-normal distribution}
\usage{
dtnorm2(n = 101, mean = 0, sd = 1, alpha = 0.05, integers = TRUE)
}
\arguments{
\item{n}{number of quantiles to r}

\item{a}{coefficient}

\item{b}{exponent}
}
\description{
Probability density of a variable following a truncated-normal distribution. This function calls msm::dtnorm internally but has different behavior.
}
\examples{

plot(dtnorm2(),type="o")
plot(dtnorm2(integer=FALSE),type="o")

}
\author{
Nikolai Klibansky
}
