% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_lin.r
\name{exp_lin}
\alias{exp_lin}
\title{Exponentially transformed linear model}
\usage{
exp_lin(x, a, b)
}
\arguments{
\item{x}{numeric vector}

\item{a}{intercept}

\item{b}{coefficient}
}
\description{
Exponentially transformed linear model exp(a+b*x)
}
\examples{
# Example

}
\author{
Nikolai Klibansky
}
