% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_pow3_norm.r
\name{fit_pow3_norm}
\alias{fit_pow3_norm}
\title{Fit power model with with three parameters and normal error to y~x data}
\usage{
fit_pow3_norm(x, y, args_optim = list())
}
\arguments{
\item{x}{independent variable}

\item{y}{dependent variable}

\item{args_optim}{Additional arguments to pass to \code{\link[stats]{optim}}}
}
\description{
Fit power model \code{\link[FishRepro]{pow3}} with three parameters, where residual error follows a normal distribution.
}
\examples{
# Fit model to batch fecundity~length data
x_ob <- data_fb_L$TL
y_ob <- data_fb_L$fb
fit <- fit_pow3_norm(x=x_ob,y=y_ob,args_optim = list(control=list(maxit=1000)))
x_pr <- seq(min(x_ob),max(x_ob),length=100)
y_pr <- do.call(pow3,c(list(x=x_pr),as.list(fit$par)))
plot(x_ob,y_ob)
points(x_pr,y_pr,type="l",lwd=2)

}
\author{
Nikolai Klibansky
}
