% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_pow_nbinom.r
\name{fit_pow_nbinom}
\alias{fit_pow_nbinom}
\title{Fit power model with two parameters and negative binomial error to y~x data.}
\usage{
fit_pow_nbinom(x, y, klim = c(0, 10), args_optim = list())
}
\arguments{
\item{x}{numeric vector. independent variable}

\item{y}{numeric vector. dependent variable}

\item{klim}{Interval to search identify starting value of dispersion parameter k. Passed to the interval argument in \code{\link[stats]{optimize}}.}

\item{args_optim}{Additional arguments to pass to \code{\link[stats]{optim}}}
}
\description{
Fit power model \code{\link[FishRepro]{pow}} with two parameters, where residual error follows a negative binomial distribution.
}
\examples{
# Fit model to batch fecundity~length data
x_ob <- data_fb_L$TL
y_ob <- data_fb_L$fb
fit <- fit_pow_nbinom(x=x_ob,y=y_ob)
x_pr <- seq(min(x_ob),max(x_ob),length=100)
y_pr <- do.call(pow,c(list(x=x_pr),as.list(fit$par[c("a","b")])))
plot(x_ob,y_ob)
points(x_pr,y_pr,type="l",lwd=2)

}
\author{
Nikolai Klibansky
}
