% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floorf.r
\name{floorf}
\alias{floorf}
\title{Compute the greatest multiple of a factor f which is less than or equal to a value x}
\usage{
floorf(x, f)
}
\arguments{
\item{x}{numeric vector}

\item{f}{factor}
}
\description{
Compute the greatest multiple of a factor f which is less than or equal to a value x
}
\examples{
\dontrun{
# Example
}

}
\author{
Nikolai Klibansky
}
