% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgs_ms.r
\name{lgs_ms}
\alias{lgs_ms}
\title{Logistic model with two parameters}
\usage{
lgs_ms(x, m, s)
}
\arguments{
\item{x}{numeric vector}

\item{m}{point of inflection parameter}

\item{s}{scale parameter}
}
\description{
Logistic model with two parameters (\eqn{y = 1/(1+exp(-(x-m)/s))})
}
\examples{
# Example
x <- 10:100
y <- lgs_ms(x=x,m=50,s=2)
y2 <- lgs_ms(x=x,m=50,s=5)
plot(x,y,type="o")
points(x,y2,type="o",col="blue")

}
\author{
Nikolai Klibansky
}
