% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgs_msa.r
\name{lgs_msa}
\alias{lgs_msa}
\title{Logistic model with three parameters}
\usage{
lgs_msa(x, m, s, a)
}
\arguments{
\item{x}{numeric vector}

\item{m}{point of inflection parameter}

\item{s}{scale parameter}

\item{a}{upper asymptote parameter}
}
\description{
Logistic model with three parameters (\eqn{y = a/(1+exp(-(x-m)/s))})
}
\examples{
# Example
x <- 10:100
y <- lgs_msa(x=x,m=50,s=2,a=0.8)
y2 <- lgs_msa(x=x,m=50,s=2,a=0.5)
plot(x,y,type="o",ylim=c(0,1))
points(x,y2,type="o",col="blue")

}
\author{
Nikolai Klibansky
}
