% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgs_msaf.r
\name{lgs_msaf}
\alias{lgs_msaf}
\title{Logistic model with four parameters}
\usage{
lgs_msaf(m, s, a, f, x)
}
\arguments{
\item{m}{point of inflection parameter}

\item{s}{scale parameter}

\item{a}{upper asymptote parameter}

\item{f}{lower asymptote parameter}

\item{x}{numeric vector}
}
\description{
Logistic model with four parameters (\eqn{y = f+(a-f)/(1+exp(-(x-m)/s))})
}
\examples{
# Example
x <- 10:100
y <- lgs_msaf(x=x,m=50,s=2,a=0.8,f=0.1)
y2 <- lgs_msaf(x=x,m=50,s=2,a=0.5,f=0.2)
plot(x,y,type="o",ylim=c(0,1))
points(x,y2,type="o",col="blue")

}
\author{
Nikolai Klibansky
}
