% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin.r
\name{lin}
\alias{lin}
\title{Linear model}
\usage{
lin(x, a, b)
}
\arguments{
\item{x}{numeric vector}

\item{a}{intercept}

\item{b}{coefficient}
}
\description{
Linear model (a+b*x)
}
\examples{
# Example

}
\author{
Nikolai Klibansky
}
