% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pla_msd.R
\name{pla_msd}
\alias{pla_msd}
\title{Plateau model with three parameters}
\usage{
pla_msd(x, m, s, d)
}
\arguments{
\item{x}{numeric vector}

\item{m}{point of inflection for left logistic function}

\item{s}{scale parameter shared by both logistic functions}

\item{d}{difference between left and right points of inflection. When fit to binary event data, average duration of events.}
}
\description{
Plateau model with three parameters (\eqn{y = 1/(1+exp(-(x-m)/s)) - 1/(1+exp(-(x-(m+d))/s))}). A plateau model is type of double logistic model where the slope of the first (left) logistic function is positive and the slope of the second (right) logistic function is negative, resulting in a domed shape. If the slopes are steep enough and the points of inflection are far enough apart, then the model will have a flat-topped section between the points of inflection.
}
\examples{
# Example
}
\author{
Nikolai Klibansky
}
