% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pla_msd_NLL.R
\name{pla_msd_NLL}
\alias{pla_msd_NLL}
\title{Objective function for fitting plateau model with three parameters}
\usage{
pla_msd_NLL(
  params,
  x,
  y,
  parLo = list(m = 1, s = 0, d = 1),
  parUp = list(m = 365, s = 100, d = 365)
)
}
\arguments{
\item{x}{numeric vector}

\item{parLo}{Lower bounds for parameters}

\item{m}{point of inflection for left logistic function}

\item{s}{scale parameter shared by both logistic functions}

\item{d}{difference between left and right points of inflection. When fit to binary event data, average duration of events.}

\item{palUp}{Upper bounds for parameters}
}
\description{
Objective function for fitting plateau model with three parameters \code{\link[FishRepro]{pla_msd}}.
}
\note{
A plateau model is type of double logistic model where the slope of the first (left) logistic function is positive and the slope of the second (right) logistic function is negative, resulting in a domed shape. If the slopes are steep enough and the points of inflection are far enough apart, then the model will have a flat-topped section between the points of inflection.
}
\examples{
# Example
dt <- FishRepro::data_SI_d_raw
dt$Start.date <- as.Date(dt$Start.date)
dt$doy <- as.numeric(format(dt$Start.date,"\%j"))
pla_msd_NLL(params=c("m"=round(min(dt$doy)+diff(range(dt$doy))*0.25),
                     "s"=1,
                     "d"=diff(range(dt$doy))/2),
            x=dt$doy,
            y=dt$SI
)
}
\author{
Nikolai Klibansky
}
