% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spawn.r
\name{plot_spawn}
\alias{plot_spawn}
\title{Plot spawning activity over time}
\usage{
plot_spawn(
  data_pop,
  data_smp = NULL,
  time_unit = "month",
  nm_spnr = "fm",
  year = "1970",
  dh_Bs = 1,
  dh_SC = 24,
  dh_HO = 12,
  dh_PF = 24,
  dh_SI = dh_HO + dh_PF,
  correct_pop = TRUE,
  filter_data_h = TRUE,
  nm_trt_plot = c("Bs", "SC", "HO", "PF", "SI"),
  args_matplot = list(),
  cols = c(Bs = rgb(0, 0, 0, 0.75), SC = rgb(1, 0.55, 0, 0.75), HO = rgb(0.33, 0.1, 0.55,
    0.75), PF = rgb(1, 0.08, 0.58, 0.75), SI = rgb(0, 0.7, 0, 0.75)),
  types = c(Bs = "o", SC = "l", HO = "l", PF = "l", SI = "l"),
  pchs = c(Bs = 16, SC = NA, HO = NA, PF = NA, SI = NA),
  ltys = c(pop = "solid", smp = "dashed")
)
}
\arguments{
\item{data_pop}{Population data returned by \code{sim_spawn}}

\item{data_smp}{Sample data returned by \code{sim_sample}}

\item{time_unit}{unit of time to use for grouping observations for plotting proportions defaults to "day", but could also be "week" or "month"}

\item{nm_spnr}{Name of column which represents numbers of spawners.}

\item{year}{Reference year for plotting data over time. Used to convert dates in data.}

\item{dh_Bs}{Duration of spawning events in hours. Spawning events are effectively instantaneous and are evident for 1 hr.}

\item{dh_SC}{Duration of spawning capable stage in hours. Even though spawning capable stage may last for months, it's best to set this to 24 hours, resulting in a correction factor of 1. (cf_SC = 24/dh_SC)}

\item{dh_HO}{Duration of hydrated oocytes (HO) in hours.}

\item{dh_PF}{Duration of post-ovulatory follicles (POF) in hours.}

\item{dh_SI}{Duration of spawning indicators in hours. Defaults to the sum of dh_HO and dh_POF}

\item{correct_pop}{Should population parameters, corrected based on dur, be plotted? If TRUE, plots of true proportions will be based on corrected data. If FALSE, the raw values will be plotted. Note that corrected values will tend to align better than uncorrected.}

\item{filter_data_h}{Should data_pop be filtered to only include dates when spawning was observed before plotting traits by hour?}

\item{nm_trt_plot}{Names of traits to plot.}

\item{args_matplot}{Additional arguments to pass to \code{\link[base]{matplot}}.}

\item{cols}{Plot colors associated with each plotted trait. Bs = spawns, SC = spawning capable, HO = hydrated oocytes, PF = spawning capable, SI = spawning indicators}

\item{types}{Plot types associated with each plotted trait. Bs = spawns, SC = spawning capable, HO = hydrated oocytes, PF = spawning capable, SI = spawning indicators}

\item{pchs}{Plotting characters associated with each plotted trait. Bs = spawns, SC = spawning capable, HO = hydrated oocytes, PF = spawning capable, SI = spawning indicators}

\item{ltys}{Line types used in plots to differentiate between population (pop) and sample (smp) data sets.}
}
\description{
Plot proportion of mature females spawning and with spawning indicators by time unit, using data produced by \code{sim_spawn} and \code{sim_sample}. Since spawning events occur instantaneously and evidence of these events (i.e. presence of HO or POF) is brief, plotting the data in a useful way can be challenging.
!!! FUNCTION UNDER CONSTRUCTION!!!
}
\examples{
\dontrun{
# Simulate spawning with defaults
set.seed(23456)
out_spawn <- sim_spawn()

# plot results with defaults
plot_spawn(data_pop=out_spawn$data_pop)

# Simulate sampling with defaults
out_sample <- sim_sample(sim_spawn_out = out_spawn)


# plot results with defaults, adding sampling data
plot_spawn(data_pop=out_spawn$data_pop,
           data_smp = out_sample$data_smp
)
}

}
\author{
Nikolai Klibansky
}
