% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow.r
\name{pow}
\alias{pow}
\title{Power model with two parameters}
\usage{
pow(x, a, b)
}
\arguments{
\item{x}{independent variable}

\item{a}{coefficient}

\item{b}{exponent}
}
\description{
Power model with two parameters (a*x^b)
}
\examples{
# Example

}
\author{
Nikolai Klibansky
}
