% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow3.r
\name{pow3}
\alias{pow3}
\title{Power model with three parameters}
\usage{
pow3(x, a, b, c)
}
\arguments{
\item{x}{independent variable}

\item{a}{intercept}

\item{b}{coefficient}

\item{c}{exponent}
}
\description{
Power model with three parameters (a+b*x^c)
}
\examples{
# Example

}
\author{
Nikolai Klibansky
}
