% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow3_nbinom_NLL.r
\name{pow3_nbinom_NLL}
\alias{pow3_nbinom_NLL}
\title{Objective function for fitting three parameter power function with negative binomial error}
\usage{
pow3_nbinom_NLL(params, x, y)
}
\arguments{
\item{params}{Named vector of parameters \code{a}, \code{b}, and \code{c} in \code{\link[FishRepro]{pow3}} function and dispersion parameter \code{k} passed to the \code{size} argument of \code{\link[stats]{dnbinom}}. \code{params} must be in order: \code{a}, \code{b}, \code{c}, \code{k}}

\item{x}{numeric vector. independent variable}

\item{y}{numeric vector. dependent variable}
}
\description{
Objective function for \code{\link[FishRepro]{pow3}} model with three parameters, when residual error of y follows a negative binomial distribution. Computes negative log likelihood value. Parameter values are input in a manner compatible with \code{\link[stats]{optim}}.
}
\examples{
x_ob <- data_fb_L$TL
y_ob <- data_fb_L$fb

fit <- lm(log(y_ob)~log(x_ob))

# Generate guesses for a, b, and c parameters
gs.a <- 1
gs.b <- exp(coef(fit)[[1]])
gs.c <- coef(fit)[[2]]

# Compute negative log-likelihood for a particular set of parameter values
pow3_nbinom_NLL(params=c("a"=gs.a,"b"=gs.b,"c"=gs.c,"k"=2),x=x_ob,y=y_ob)

}
\author{
Nikolai Klibansky
}
