% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow_nbinom_NLL.r
\name{pow_nbinom_NLL}
\alias{pow_nbinom_NLL}
\title{Objective function for fitting power function with two parameters and negative binomial error}
\usage{
pow_nbinom_NLL(params, x, y)
}
\arguments{
\item{params}{Named vector of parameters \code{a} and \code{b} in \code{\link[FishRepro]{pow}} function and dispersion parameter \code{k} passed to the \code{size} argument of \code{\link[stats]{dnbinom}}. \code{params} must be in order: \code{a}, \code{b}, \code{k}}

\item{x}{numeric vector. independent variable}

\item{y}{numeric vector. dependent variable}
}
\description{
Objective function for \code{\link[FishRepro]{pow}} model with two parameters, when residual error follows a negative binomial distribution. Computes negative log likelihood value. Parameter values are input in a manner compatible with \code{\link[stats]{optim}}.
}
\examples{
x_ob <- data_fb_L$TL
y_ob <- data_fb_L$fb

fit <- lm(log(y_ob)~log(x_ob))

# Generate guesses for a and b parameters based on linear fit of transformed data
gs.a <- exp(coef(fit)[[1]])
gs.b <- coef(fit)[[2]]

# Compute negative log-likelihood for a particular set of parameter values
pow_nbinom_NLL(params=c("a"=gs.a,"b"=gs.b,"k"=2),x=x_ob,y=y_ob)

}
\author{
Nikolai Klibansky
}
