% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_sample.r
\name{sim_sample}
\alias{sim_sample}
\title{Simulate sampling of a population of fish}
\usage{
sim_sample(
  sim_spawn_out = NULL,
  d_sim = 365,
  dates_smp_n = 30,
  Sn0_m = 4,
  Sn0_d = 1,
  Sn1_m = 10,
  Sn1_d = 31,
  CrSn0_m = 5,
  CrSn0_d = 1,
  CrSn1_m = 9,
  CrSn1_d = 30,
  Ll_n = 4,
  Ll_cpue_mn = 0.698,
  Ll_cpue_sd = 0.537,
  Ll_TL_mean = 582.2029,
  Ll_TL_sd = 100.6291,
  sex_rat = 0.5,
  n_Lbinpsc = 1,
  sampling_annual = 2,
  sampling_pattern = 4,
  sampling_sizeselect = 4,
  weather_ratio = 0.135,
  args_diel_sample = list(freq = 1, pattern = "day")
)
}
\arguments{
\item{sim_spawn_out}{Object returned by \code{\link[FishRepro]{sim_spawn}}}

\item{dates_smp_n}{Number of dates on which to sample the population}

\item{Sn0_m}{Start of known spawning season: month}

\item{Sn0_d}{Start of known spawning season: day}

\item{Sn1_m}{End of known spawning season: month}

\item{Sn1_d}{End of known spawning season: day}

\item{CrSn0_m}{Start of cruise (sampling) season: month}

\item{CrSn0_d}{Start of cruise (sampling) season: day}

\item{CrSn1_m}{End of cruise (sampling) season: month}

\item{CrSn1_d}{End of cruise (sampling) season: day}

\item{Ll_n}{Number of longlines to pull at each sampling event. Multiplied by empirical mean cpue to allow simulated manipulation of catch rates.}

\item{Ll_cpue_mn}{Mean longline catch rate (both sexes) CPUE (number of fish per longline set; 100 hooks per 1 hour set)}

\item{Ll_cpue_sd}{Standard deviation of longline catch rate (both sexes) CPUE (number of fish per longline set. 100 hooks, 1 hour set)}

\item{Ll_TL_mean}{Mean total length of fish in longline catches}

\item{Ll_TL_sd}{Standard deviation of fish in longline catches}

\item{sex_rat}{Sex ratio in population}

\item{n_Lbinpsc}{Number of females to select per size class per longline. Only used when sampling_sizeselect is 4 or 5.}

\item{sampling_annual}{Annual sampling type: 1=sample all year, 2=sample during known spawning months only, 3=during cruise season}

\item{sampling_pattern}{Temporal sampling pattern: 2=random, 3=standard survey, 4=directed research}

\item{sampling_sizeselect}{Size-based sampling type. Types 1 and 2 both determine the number of females to sample as the number of length bins in the model (n_Lbin). Type 1 selects a perfectly uniform size distribution (one female per length bin), while type 2 selects n_Lbin  lengths in a random uniform manner. Types 4 and 5 both determine the number of females to sample based on empirical catch data by making a random draw from a normal distribution with mean = Ll_cpue_mn*Ll_n and sd = Ll_cpue_sd. When type=4 the function samples randomly from a normal size-frequency distribution with mean = Ll_TL_mean and sd = Ll_TL_sd. If sampling_sizeselect==5, sample from a uniform size-frequency distribution.}

\item{weather_ratio}{unfishable days/fishable days  (i.e. 0 indicates all days are fishable)}

\item{args_diel_sample}{Arguments passed to \code{\link[FishRepro]{diel_sample}} function.}
}
\description{
Simulate sampling of a population of fish
!!! FUNCTION UNDER CONSTRUCTION!!!
}
\examples{
\dontrun{
# Simulate spawning with defaults
set.seed(23456)
out_spawn <- sim_spawn()

# Simulate sampling with defaults
out_sample <- sim_sample(sim_spawn_out = out_spawn)

# Set dh_HO and dh_PF to match values in \code{sim_spawn}
par(mfrow=c(3,1),mar=c(3,3,1,1),mgp=c(1,0.2,0),tck=-0.01)
plot_spawn(data_pop=  out_spawn$data_pop,
           data_smp = out_sample$data_smp,
           time_unit = "month",
           dh_HO = 6,
           dh_PF = 8
)
plot_spawn(data_pop=out_spawn$data_pop,
           data_smp = out_sample$data_smp,
           time_unit = "week",
           dh_HO = 6,
           dh_PF = 8)
plot_spawn(data_pop=out_spawn$data_pop,
           data_smp = out_sample$data_smp,
           time_unit = "day",
           dh_HO = 6,
           dh_PF = 8)
par(mfrow=c(1,1),mar=c(3,3,1,1),mgp=c(1,0.2,0),tck=-0.01)
plot_spawn(data_pop=out_spawn$data_pop,
           data_smp = out_sample$data_smp,
           time_unit = "hour",
           dh_HO = 6,
           dh_PF = 8,
           nm_trt_plot  = c("Bs","HO","PF"))

# Simulate sampling with some minor improvements to sampling
out_sample <- sim_sample(sim_spawn_out = out_spawn,
                         args_diel_sample=list(freq=1,pattern="all")
)

# Set dh_HO and dh_PF to match values in \code{sim_spawn}
par(mfrow=c(3,1),mar=c(3,3,1,1),mgp=c(1,0.2,0),tck=-0.01)
plot_spawn(data_pop=  out_spawn$data_pop,
           data_smp = out_sample$data_smp,
           time_unit = "month",
           dh_HO = 6,
           dh_PF = 8
)
plot_spawn(data_pop=out_spawn$data_pop,
           data_smp = out_sample$data_smp,
           time_unit = "week",
           dh_HO = 6,
           dh_PF = 8)
plot_spawn(data_pop=out_spawn$data_pop,
           data_smp = out_sample$data_smp,
           time_unit = "day",
           dh_HO = 6,
           dh_PF = 8)
par(mfrow=c(1,1),mar=c(3,3,1,1),mgp=c(1,0.2,0),tck=-0.01)
plot_spawn(data_pop=out_spawn$data_pop,
           data_smp = out_sample$data_smp,
           time_unit = "hour",
           dh_HO = 6,
           dh_PF = 8,
           nm_trt_plot  = c("Bs","HO","PF"))

}

}
\author{
Nikolai Klibansky
}
