% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_spawn.r
\name{sim_spawn}
\alias{sim_spawn}
\title{Simulate a spawning population of fish}
\usage{
sim_spawn(
  Linf = 911.36,
  K = 0.24,
  t0 = -0.33,
  Lmin = 200,
  Lbinw = 20,
  a50 = 2,
  a_lw = 1.65e-05,
  b_lw = 2.99,
  d_sim = 365,
  j.mn.SI = 156,
  j.sd.SI = 38.62560942,
  Ps.mn = 0.389112903,
  Ps.m = 580.1673777,
  Ps.s = 66.92958968,
  Ps.a = 0.606122263,
  Ps.f = 0.213780242,
  d.mn.Sn = 78,
  h.mn.HO = 10.19994175,
  h.sd.HO = 1.770091189,
  h.mn.spn = 16.38637712,
  h.sd.spn = 4.863383191,
  d.mn.HO = 6,
  d.mn.PF = 8,
  SI.dur = 14,
  SI_cf = 1.714285714,
  fb.a = 0.407975854,
  fb.b = 3.03e-08,
  fb.c = 4.774001527,
  fb.k = 2.618605045,
  fb.thresh = 375,
  fb.min = 58935,
  Sn0.m = 4,
  Sn0.d = 1,
  Sn1.m = 10,
  Sn1.d = 31,
  nm_fn_fb = "pow3",
  first_batch_error = TRUE,
  diel_spawn_error = TRUE,
  POF_TempDep = FALSE,
  fb_size_error = TRUE,
  SI_size_error = TRUE,
  covmat_lgs_msaf = covmat_SI_L,
  BI_size_independent = FALSE
)
}
\arguments{
\item{Linf}{length infinity}

\item{K}{growth coefficient}

\item{t0}{age (time) at length zero}

\item{Lmin}{minimum length of fish (same units as Linf)}

\item{Lbinw}{width of length bins (same units as Linf)}

\item{a50}{Age at maturity using knife edge maturity function (analogous to age at 50 percent maturity from logistic function)}

\item{a_lw}{a parameter in length weight equation (W = aL^b)}

\item{b_lw}{b parameter in length weight equation (W = aL^b)}

\item{d_sim}{number of days in the simulation}

\item{j.mn.SI}{Mean appearance of spawning markers during the year (Julian day)}

\item{j.sd.SI}{StDev appearance of spawning markers during the year (Julian day)}

\item{Ps.mn}{Mean (size-independent) spawning fraction}

\item{Ps.m}{spawning fraction~length parameter m (point of inflection)}

\item{Ps.s}{spawning fraction~length parameter s (scale parameter)}

\item{Ps.a}{spawning fraction~length parameter a (upper asymptote)}

\item{Ps.f}{spawning fraction~length parameter f ( lower asymptote (i.e. floor))}

\item{d.mn.Sn}{Mean duration of spawning period (days)}

\item{h.mn.HO}{Mean time (hour of the day) that HO become evident}

\item{h.sd.HO}{StDev time (hour of the day) that HO become evident}

\item{h.mn.spn}{Mean time (hour of the day) that females spawn (POF become evident)}

\item{h.sd.spn}{StDev time (hour of the day) that females spawn (POF become evident)}

\item{d.mn.HO}{mean duration of hydrated oocytes (in hours)}

\item{d.mn.PF}{mean duration in hours of post ovulatory follicles (in hours)}

\item{SI.dur}{mean duration of spawning indicators (in hours), used when spawning indicator is the union of two events (i.e. SI presence = presence of either HO or POF)}

\item{SI_cf}{spawning indicator correction factor: 24/SI.dur}

\item{fb.a}{Batch fecundity-length intercept parameter}

\item{fb.b}{Batch fecundity-length coefficient parameter}

\item{fb.c}{Batch fecundity-length exponent parameter}

\item{fb.k}{Batch fecundity-length k parameter for negative binomial error distribution}

\item{fb.thresh}{Threshold length (TL mm), below which fecundity is fixed at the observed value}

\item{fb.min}{Fecundity (observed) for fish < threshold length}

\item{Sn0.m}{Start of known spawning season: month}

\item{Sn0.d}{Start of known spawning season: day}

\item{Sn1.m}{End of known spawning season: month}

\item{Sn1.d}{End of known spawning season: day}

\item{nm_fn_fb}{Name of function used to compute batch fecundity}

\item{first_batch_error}{Include an error distribution around the date that the first batch of the season is spawned?}

\item{diel_spawn_error}{Include an error distribution around the time of day that batches are spawned?}

\item{POF_TempDep}{Should POF duration be calculated as a function of water temperature based on empirical data?}

\item{fb_size_error}{Should error be included around the batch fecundity-length function?}

\item{SI_size_error}{Should error be included around the spawning fraction-length function?}

\item{covmat_lgs_msaf}{Covariance matrix for parameters in four parameter logistic model fit to spawning fraction-length data. Only used when SI_size_error=TRUE.}

\item{BI_size_independent}{Should batch interval be independent of body size?}
}
\description{
Simulate a spawning population of fish to generate fecundity data.
}
\examples{
\dontrun{
# Write example here
}

}
\author{
Nikolai Klibansky
}
